Rename variables.rs to constants.rs
